Adding methods to convert char sequences to integers and vice versa
We should explore using 2 bytes as a pointer instead of a 4 int as char
representation! This would reduce the directory size from 20 to 18.
Also, if we know the max length of a name to be 15 chars, why is the
16th char have to be a null terminating character?

Also, see that the directory method now reflects the aassignment's
specs.