New functions
methods to convert sectors for the data blocks into useful information
for the maps. For example, the 1st data block is actually soemthing like
sector 1000. Teh ConvertDataMapToAbsolute does this math.
Conversely, the 2000th sector is something like the 1000 data block. In
order to get information on the 2000th sector, which is a data block,
you will need to use the ConvertDataSectorToMapSectorNum method.

Just remember that the data map starts at 0 itself but it's mapping
starts at the 1000th sector. Conversions are needed!

Also, conversions will be for inodes..... IM LOST! I might need to blow
these methods up and start over!